moving search components from general components directory to search/components
