+++ packages/openneuro-app/src/scripts/search/__tests__/NuerobagelSearch.spec.tsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { fireEvent, render, screen } from "@testing-library/react"
-import { NeurobagelSearch } from "../NeurobagelSearch"
+import { NeurobagelSearch } from "../components/NeurobagelSearch"
 
 describe("NeurobagelSearch component", () => {
   it("? toggle can be clicked", async () => {
