+++ packages/openneuro-app/src/scripts/search/components/FacetBlockContainerExample.tsx
