+++ packages/openneuro-app/src/scripts/search/components/ModalityHeader.tsx
