+++ packages/openneuro-app/src/scripts/search/components/SearchResultItem.tsx
@@ -7,7 +7,7 @@ import { Icon } from "../../components/icon/Icon"
 import { useCookies } from "react-cookie"
 import { getProfile } from "../../authentication/profile"
 import { useUser } from "../../queries/user"
-import "./search-result.scss"
+import "../scss/search-result.scss"
 import activityPulseIcon from "../../../assets/activity-icon.png"
 import { hasEditPermissions } from "../../authentication/profile"
 import { ModalityHexagon } from "../../components/modality-cube/ModalityHexagon"
