+++ packages/openneuro-app/src/scripts/search/inputs/index.ts
@@ -2,7 +2,7 @@ import KeywordInput from "./keyword-input"
 import AllDatasetsToggle from "./admin-allDatasets-toggle"
 import ModalitySelect from "./modality-select"
 import InitiativeSelect from "./initiative-select"
-import ShowDatasetRadios from "./show-datasets-radios"
+import ShowDatasetRadios from "../components/show-datasets-radios"
 import AgeRangeInput from "./age-range-input"
 import SubjectCountRangeInput from "./subject-count-range-input"
 import DatasetTypeSelect from "./dataset-type-select"
