+++ packages/openneuro-app/src/scripts/search/scss/filters-block.scss
