+++ packages/openneuro-app/src/scripts/search/scss/search-result.scss
