+++ packages/openneuro-app/src/scripts/search/scss/sliding-radio-group.scss
@@ -1,4 +1,3 @@
-
 @import '../../scss/variables';
 
 /*
@@ -6,27 +5,31 @@
   Apply box-sizing globally or to the component root for consistency
   -------------------------------------------------------------
 */
-
+*,
+*::before,
+*::after {
+  box-sizing: border-box;
+}
 
 /*
   -------------------------------------------------------------
-  Root container for the ShowDatasetRadios component
-  All component-specific styles are nested here to prevent conflicts.
+  Root container for the general SlidingRadioGroup component
   -------------------------------------------------------------
 */
-.on-show-dataset-radios-root {
-  *, *::before, *::after {
-  box-sizing: border-box;
-}
+.sliding-radio-group-root {
   &.btn-group-wrapper.facet-radio.show-dataset-radios-container {
     position: relative;
-    margin: 0 0 10px; 
+    margin: 0 0 10px;
     background-color: #f6f6f6;
     border: 1px solid #eee;
     border-radius: 4px;
     overflow: hidden;
+    min-height: 40px;
+
     display: flex;
-    display: inline-block;
+    flex-direction: row;
+    align-items: stretch;
+    justify-content: flex-start;
 
     &:after {
       content: none;
@@ -40,11 +43,11 @@
 
   /*
     -------------------------------------------------------------
-    Styles for .dataset-filter-radio adapted for button groups
+    Styles for .dataset-filter-radio (individual radio button wrapper)
     -------------------------------------------------------------
   */
-  .dataset-filter-radio { 
-    margin: 0; 
+  .dataset-filter-radio {
+    margin: 0;
 
     [type='radio']:checked,
     [type='radio']:not(:checked) {
@@ -60,54 +63,56 @@
       display: flex;
       align-items: center;
       justify-content: center;
-      height: 100%; 
+      height: 100%;
       white-space: nowrap;
-      padding: 6px 12px; 
+      padding: 6px 12px;
 
-      border-right: 1px solid #ddd; 
+      border-right: 1px solid #ddd;
 
       &::before,
       &::after {
         display: none;
       }
     }
+
     &:last-child > label,
     label.is-active {
-      border-right: none; 
+      border-right: none;
     }
   }
 
   /*
     -------------------------------------------------------------
-    Styles for the inner group of radio buttons and the sliding highlight
+    Styles for the inner group of radio buttons container
     -------------------------------------------------------------
   */
   .show-dataset-radios-group {
     display: flex;
-    flex-direction: row; 
+    flex-direction: row;
     justify-content: flex-start;
-    align-items: stretch; 
+    align-items: stretch;
     z-index: 1;
     position: relative;
     width: 100%;
   }
 
-  
   .sliding-highlight {
     position: absolute;
     background-color: var(--current-theme-primary);
     border-radius: 4px;
-    transition: left 0.3s ease, width 0.3s ease, top 0.3s ease, height 0.3s ease, opacity 0.3s ease;
+
+    transition: left 0.3s ease, width 0.3s ease, top 0.3s ease, height 0.3s ease,
+      opacity 0.3s ease;
     z-index: 0;
     opacity: 0;
   }
 
   /*
     -------------------------------------------------------------
-    Specific button states within this component
+    Specific button states within this component (text color changes)
     -------------------------------------------------------------
   */
-  .dataset-filter-radio { 
+  .dataset-filter-radio {
     > label {
       color: var(--current-theme-primary);
       background-color: transparent;
@@ -122,41 +127,5 @@
     }
   }
 
-  /*
-    -------------------------------------------------------------
-    Responsive Adjustments (Mobile first for column layout)
-    -------------------------------------------------------------
-  */
-  @media (max-width: 787px) {
-    &.btn-group-wrapper.facet-radio.show-dataset-radios-container {
-      flex-direction: column;
-      height: auto;
-      justify-content: flex-start;
-      align-items: flex-start;
-    }
 
-    .show-dataset-radios-group {
-      flex-direction: column;
-      width: 100%;
-      height: auto;
-      align-items: stretch;
-    }
-
-    .sliding-highlight {
-      width: 100% !important;
-      left: 0 !important;
-    }
-
-    .dataset-filter-radio { 
-      > label {
-        height: 50px;
-        min-width: 100%;
-        border-bottom: 1px solid var(--current-theme-primary-ultralight);
-        border-right: none; 
-        &:last-child {
-          border-bottom: none;
-        }
-      }
-    }
-  }
 }
